pyproject.toml
[project]
name = "email_client"
version = "0.1.0"
description = "Email search and management tool using MCP protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.1.2",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
email-client = "email_client:main"